requirements.txt
Flask==2.3.3
Pillow==10.0.1
python-socketio==5.8.0
eventlet==0.33.3 